requirements.txt
aiogram>=3.4,<4.0
requests
aiohttp>=3.8.0
aiosqlite>=0.17.0
sqlalchemy>=2.0.0
python-dotenv>=0.20.0
openpyxl>=3.1.5
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.10.0
webdriver-manager>=4.0.0
python-multipart
brotli>=1.0.0
greenlet>=3.2.4